Retire the move-method gci e2e suite, already covered in CI
Both tests in `gci/gciMoveMethod.e2e.test.ts` were duplicates of
`refactoring/__tests__/refactoringMoveMethod.integration.test.ts`, which drives
the same query builders and parsers over the same fixture shape and asserts
strictly more — and, unlike the gci project, actually runs in CI across the
release matrix in both the bare and plugin-installed passes.

The one edge the sibling did not cover was a preview where EVERY selected
method is non-movable (the gci suite previewed the super-sender alone, while the
integration test only saw it alongside two movable selectors), so that case
lands there as its own test before the duplicate goes away.
